[CAN-7] refactor: 캘린더 헤더 컴포넌트 분리
